Add current secret key base to database migrations
For better or worse, we are generating a new secret key base with each
run of the main task and storing it in an environment variable
accessible to the application owner. Subsequent configuration tasks
either cannot access this value or they can, but use the previous value
when the tasks are re-run. Not certain which is the case. In either event, explicitly
providing the current value to the task fixes the problem.