Refine post-install configuration tasks
Use rails rather than rake (to avoid some errors that suddenly started
cropping up) and explicitly pass db password as an environment
variable to the tasks (rather than storing it as a global environment
variable on the server). Finally, store database password in the
application user's environment for safer keeping.