Run post-install configuration tasks as app user
After a point, further configuration needs to be done as the application
user, rather than the ansible/login/deploy user. This requires creating
the app user as part of the main task (to make sure it always happens)
and installing rails specifically for the app user (in addition to the a/l/d
user).