Stop setting global environment vars
The way we were adding the bundler and gems paths to the everyones path
environment var was faulty (it would add them again each time this play
was run). When we went in to fix it, it wasn't clear that this was a
necessary or beneficial thing to do anyway, so we're 86'ing it for now.